requirements.txt
numpy
tensorflow_gpu
keras
